Reduce branches: use a table to convert ascii to digit values.
In paticular the hex numbers with potential lower and upper cases
create a lot of branches just to figure out if a value is in range.
It is cheaper to just have a look-up table that maps ascii chars to
the respective digit value.

Ascii chars that are not digits come in two categories: things that
are not digits and should trigger stop parsing the number.
For our Verilog-inspired numbers, we have the underscore character
that can be part of the number (as digit separator), but does not
represent a digit - but parsing should continue.

To be able to determine if we use a character just with a single
comparison, all ASCII characters in the first category are encoded
as a large number that is not a valid digit (for aestetic reasons,
let's use 99 here).

The underscore is encoded as -1, so _less_ than the smallest
digit 0 in all bases.

That way we can do a single compare to determine if we should
continue parsing the number.

   while (lookup_table[character] < desired_number_base) ...

That way, we get to about 600MiB/s parse speed.